build: move runPostMethod and auxiliary functions to Client/postMethod.cpp refactor: move runPostMethod to Client class from Reponse class feat: add setStatusMessage to Reponse class
